+++ osidb/models.py
@@ -2,6 +2,7 @@
 draft model for end to end testing
 """
 import logging
+import re
 import uuid
 from decimal import Decimal
 from typing import Union
@@ -808,6 +809,35 @@ class Flaw(WorkflowModel, TrackingMixin, NullStrFieldsMixin, AlertMixin):
         old_flaw = Flaw.objects.get(pk=self.pk)
         if not old_flaw.embargoed and self.is_embargoed:
             raise ValidationError("Embargoing a public flaw is futile")
+
+    def _validate_cwe_format(self):
+        """
+        Check if CWE string is well formated
+        """
+        cwe_data = self.cwe_id
+        # First, remove possible [auto] suffix from the CWE entry
+        # [auto] suffix means value was assigned by a script during mass update
+        if len(cwe_data) > 6 and cwe_data.endswith("[auto]"):
+            cwe_data = cwe_data[:-6]
+
+        # Then split data on arrows ->, later we will parse the elements separately
+        arrow_count = len(re.findall("->", cwe_data))
+        parsed_elements = list(filter(None, cwe_data.split("->")))
+
+        # Ensure number of elements is one bigger then count of arrows, to catch
+        # stuff like: CWE-123->
+        if len(parsed_elements) > 0 and len(parsed_elements) != (arrow_count + 1):
+            raise ValidationError(
+                "CWE IDs is not well formated. Incorrect number of -> in CWE field."
+            )
+
+        # Ensure each element is well formed, i.e. one of:
+        #   * CWE-123
+        #   * (CWE-123)
+        #   * (CWE-123|CWE-456)
+        for element in parsed_elements:
+            if not re.match(r"^(CWE-[0-9]+|\(CWE-[0-9]+(\|CWE-[0-9]+)*\))$", element):
+                raise ValidationError("CWE IDs is not well formated.")
 
     # TODO this needs to be refactored
     # but it makes sense only when we are capable of write actions